requirements.txt
absl-py==2.2.2
accelerate==1.7.0
astunparse==1.6.3
bitsandbytes==0.42.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.5.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
hf-xet==1.1.2
huggingface-hub==0.32.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
opt-einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
pip==24.3.1
protobuf==5.29.4
psutil==7.0.0
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scipy==1.15.3
setuptools==80.8.0
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing-extensions==4.13.2
urllib3==2.4.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
